Move out global config to lbrytv config package as a part of monorepo conversion
