+++ internal/monitor/monitor_test.go
@@ -3,7 +3,7 @@ package monitor
 import (
 	"testing"
 
-	"github.com/lbryio/lbrytv/config"
+	"github.com/lbryio/lbrytv/apps/lbrytv/config"
 
 	"github.com/sirupsen/logrus"
 	"github.com/sirupsen/logrus/hooks/test"
