+++ internal/test/test.go
@@ -10,7 +10,7 @@ import (
 	"regexp"
 	"testing"
 
-	"github.com/lbryio/lbrytv/config"
+	"github.com/lbryio/lbrytv/apps/lbrytv/config"
 
 	"github.com/nsf/jsondiff"
 	"github.com/ybbus/jsonrpc"
