+++ main.go
@@ -5,8 +5,8 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/lbryio/lbrytv/apps/lbrytv/config"
 	"github.com/lbryio/lbrytv/cmd"
-	"github.com/lbryio/lbrytv/config"
 	"github.com/lbryio/lbrytv/internal/reflection"
 	"github.com/lbryio/lbrytv/internal/storage"
 )
